SH: FilesExtractor: size_t -> uint32_t

+++ trunk/game-specific/Silent Hill/FilesExtractor/crc.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "Common.h"
 
-uint16_t crc16(const void* buf, size_t len, uint16_t crc = 0);
-uint32_t crc32(const void* buf, size_t len, uint32_t crc = 0);
+uint16_t crc16(const void* buf, uint32_t len, uint16_t crc = 0);
+uint32_t crc32(const void* buf, uint32_t len, uint32_t crc = 0);
